==> 0a-execution_script.ipynb <==
# run_pipeline.py
import papermill as pm
import sys
import os
import shutil

# Remove and recreate output folder
if os.path.exists("output"):
    shutil.rmtree("output")
os.makedirs("output", exist_ok=True)

notebooks =[
 '1-SP500MEMBERSHIPBUILDER.ipynb', 
 '1a-SP500_Join_Exit_Date.ipynb', 
 '2a-Incremental_price_fetch.ipynb', 
 '3-adjusted_All_Prices_OHLC.ipynb', 
 '4-ATR20_adjusted_All_Prices.ipynb', 
 '4a-Parameter_Sweep_ATRs.ipynb', 
 '5-100D_MA_adjusted_all_prices.ipynb', 
 '6-90Day_jump_filter_adjusted_all_prices.ipynb', 
 '7-90Day_exp_regression_adjusted_all_prices.ipynb', 
 '7a-Parameter_Sweep_multiple-regressions.ipynb', 
 '8-SPY_200DMA_market_regime.ipynb', 
 '9-build_ranking_dataset.ipynb', 
 '9a-multiple-ranking_datasets.ipynb', 
 '10-build_100MA_filtered_universe.ipynb', 
 '10a-multiple_100MA_filtered.ipynb', 
 '11-build_90D_jump_filtered_univers.ipynb', 
 '11a-mulitple_jump_filtered_universe.ipynb', 
 '12-trim_to_sp500.ipynb', 
 '12a-multiple-trim_sp500.ipynb', 
 '13-match_trade_generator.ipynb',
 '13-regime_pct_cap_trading_engine.ipynb', 
 '13a-strategy_vs_spy.ipynb', 
 '13a-sweep-trading_engine.ipynb', 
 '13b-wfo.ipynb', 
 '13c-wfo_vs_spy.ipynb', 
 '13d-wfo_monte_carlo.ipynb', 
 '13e-rankings_vs_trades.ipynb',  
 '15a-weekly_report_with_rankings.ipynb', 
 '15b-cash-monitoring.ipynb', 
 '15c-cashwithraw.ipynb', 
 '15d-position_too_big.ipynb', 
 '16-Monte-Carlo.ipynb', 
 '17-robustness_suite.ipynb',
 '17c-industry_standard_cost_slippage.ipynb', 
 '18_risk_adjusted_full_report.ipynb', 
 '23-robustness_test.ipynb', 
 '24-crash_test.ipynb', 
 '25-return_analysis.ipynb', 
 '25a-cashflow.ipynb', 
 '25b-monthly_profit.ipynb', '25c-convexity_test.ipynb', 
 '25d-Monthly-components.ipynb', '25d-min_cash.ipynb', 
]
failed = []

for nb in notebooks:
    try:
        print(f"\n{'='*50}")
        print(f"Running: {nb}")
        print('='*50)
        
        pm.execute_notebook(nb, f"output/{nb}", kernel_name="python3")
        
        print(f"✓ Completed: {nb}")
        
    except Exception as e:
        print(f"✗ Failed: {nb}")
        print(f"  Error: {e}")
        failed.append(nb)
        
        # Option: stop on first failure
        # sys.exit(1)

print(f"\n{'='*50}")
print(f"Summary: {len(notebooks) - len(failed)}/{len(notebooks)} succeeded")
if failed:
    print(f"Failed: {failed}")
